requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database dependencies
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
pymysql==1.1.0
aiosqlite==0.19.0

# Dashboard dependencies
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3

# HTTP client and utilities
requests==2.31.0
python-multipart==0.0.6

# Date/time handling
pytz==2023.3

# Data validation
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0